// libs/features/customer-information/src/lib/shared/models/benefits/marriage-details.ts
/**
 * Copyright GOSI. All Rights Reserved.
 * This software is the proprietary information of GOSI.
 * Use is subject to license terms.
 */
import { BilingualText, GosiCalendar } from '@gosi-ui/core';

export class MarriageDetails {
  startDate: GosiCalendar;
  endDate: GosiCalendar;
  event: BilingualText;
}